test: 로또 머신에 사용자 금액 입력 예외 테스트
